ios-app/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
pillow
google-generativeai